Backward compatibility changes
