+++ sdk/model/search_result_workflow_schedule_execution_model.go
@@ -10,6 +10,6 @@
 package model
 
 type SearchResultWorkflowSchedule struct {
-	Results   []WorkflowScheduleExecution `json:"results,omitempty"`
-	TotalHits int64                       `json:"totalHits,omitempty"`
+	Results   []WorkflowScheduleExecutionModel `json:"results,omitempty"`
+	TotalHits int64                            `json:"totalHits,omitempty"`
 }
