+++ sdk/model/tag.go
@@ -5,9 +5,3 @@ type Tag struct {
 	Type_ string `json:"type,omitempty"`
 	Value string `json:"value,omitempty"`
 }
-
-type TagString struct {
-	Key   string `json:"key,omitempty"`
-	Type_ string `json:"type,omitempty"`
-	Value string `json:"value,omitempty"`
-}
